Refactor TilemapGrid into modular hooks and helpers
