+++ components/tilemap-grid/constants.ts
@@ -1,0 +1,3 @@
+import { type EnemyKind } from "../../lib/enemies/registry";
+
+export const TORCH_CARRIER_ENEMIES = new Set<EnemyKind>(["goblin"]);
